smart_farming_app/requirements.txt
Flask
pandas
numpy
scikit-learn
joblib
Pillow
reportlab